feat(domain.validators): String Validator and the 'validators' package are included
